+++ processing-library/ton/src/de/hfkbremen/ton/DSPNodeInput.java
@@ -1,5 +1,0 @@
-package de.hfkbremen.ton;
-
-public interface DSPNodeInput {
-    void input(float pSignal);
-}