+++ processing-library/ton/src/de/hfkbremen/ton/DSPNodeOutput.java
@@ -1,5 +1,0 @@
-package de.hfkbremen.ton;
-
-public interface DSPNodeOutput {
-    float output();
-}