+++ processing-library/ton/src/ton/AudioBufferRenderer.java
@@ -1,0 +1,6 @@
+package ton;
+
+public interface AudioBufferRenderer {
+
+    void audioblock(float[][] pOutputSamples, float[][] pInputSamples);
+}