+++ src/de/hfkbremen/ton/AudioBufferRenderer.java
@@ -1,6 +1,0 @@
-package de.hfkbremen.ton;
-
-public interface AudioBufferRenderer {
-
-    void audioblock(float[][] pOutputSamples, float[][] pInputSamples);
-}