+++ src/de/hfkbremen/ton/BeatListener.java
@@ -1,6 +1,0 @@
-package de.hfkbremen.ton;
-
-public interface BeatListener {
-
-    void beat(int pBeatCount);
-}