+++ src/de/hfkbremen/ton/DSPNodeProcess.java
@@ -1,6 +1,0 @@
-package de.hfkbremen.ton;
-
-public interface DSPNodeProcess {
-
-    float process(float pSignal);
-}