+++ src/ton/BeatListener.java
@@ -1,0 +1,6 @@
+package ton;
+
+public interface BeatListener {
+
+    void beat(int pBeatCount);
+}