+++ src/ton/DSPNodeInput.java
@@ -1,0 +1,5 @@
+package ton;
+
+public interface DSPNodeInput {
+    void input(float pSignal);
+}