+++ src/ton/DSPNodeOutput.java
@@ -1,0 +1,5 @@
+package ton;
+
+public interface DSPNodeOutput {
+    float output();
+}