+++ src/ton/DSPNodeProcess.java
@@ -1,0 +1,6 @@
+package ton;
+
+public interface DSPNodeProcess {
+
+    float process(float pSignal);
+}